+++ KeySwitchManager/Sources/Runtime/Applications/CLI/Sources/Commands/ExportDawArticulation.cs
@@ -2,8 +2,9 @@ using CommandLine;
 
 using KeySwitchManager.Applications.CLI.Views;
 using KeySwitchManager.Applications.Standalone.Core.KeySwitches.Commons;
-using KeySwitchManager.Applications.Standalone.Core.KeySwitches.Controllers;
+using KeySwitchManager.Applications.Standalone.Core.KeySwitches.Extensions.Controllers;
 using KeySwitchManager.Applications.Standalone.Core.KeySwitches.Helpers;
+using KeySwitchManager.Controllers.KeySwitches;
 using KeySwitchManager.Domain.KeySwitches.Models.Values;
 using KeySwitchManager.Presenters.KeySwitches;
